15-greeting-agent/pyproject.toml
[project]
name = "15-greeting-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.0.10",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.0",
    "pyttsx3>=2.98",
    "speechrecognition>=3.14.2",
]
